split out JSPath
