+++ src/PostgREST/ApiRequest.hs
@@ -45,19 +45,20 @@ import Web.Cookie                (parseCookiesText)
 
 import Data.Ranged.Boundaries
 
-import PostgREST.ContentTypes     (ContentType (..),
-                                   decodeContentType, toMime)
+import PostgREST.ContentTypes          (ContentType (..),
+                                        decodeContentType, toMime)
 import PostgREST.DbStructureTypes
-import PostgREST.Error            (ApiRequestError (..))
-import PostgREST.Parsers          (pRequestColumns)
+import PostgREST.Error                 (ApiRequestError (..))
+import PostgREST.Parsers               (pRequestColumns)
 import PostgREST.Preferences
-import PostgREST.Private.QueryFragment (operators, ftsOperators)
-import PostgREST.RangeQuery       (NonnegRange, allRange, rangeGeq,
-                                   rangeLimit, rangeOffset,
-                                   rangeRequested, restrictRange)
+import PostgREST.Private.QueryFragment (ftsOperators, operators)
+import PostgREST.RangeQuery            (NonnegRange, allRange,
+                                        rangeGeq, rangeLimit,
+                                        rangeOffset, rangeRequested,
+                                        restrictRange)
 import PostgREST.Types
-import Protolude                  hiding (head, toS)
-import Protolude.Conv             (toS)
+import Protolude                       hiding (head, toS)
+import Protolude.Conv                  (toS)
 
 type RequestBody = BL.ByteString
