+++ test/Feature/AuthSpec.hs
@@ -7,8 +7,8 @@ import Test.Hspec
 import Test.Hspec.Wai
 import Test.Hspec.Wai.JSON
 
-import PostgREST.PgVersions            (PgVersion, pgVersion112)
-import Protolude                  hiding (get)
+import PostgREST.PgVersions (PgVersion, pgVersion112)
+import Protolude            hiding (get)
 import SpecHelper
 
 spec :: PgVersion -> SpecWith ((), Application)
