+++ test/SpecHelper.hs
@@ -23,7 +23,7 @@ import Test.Hspec.Wai
 import Text.Heredoc
 
 import PostgREST.Config (AppConfig (..), LogLevel (..), parseSecret)
-import PostgREST.Types  (JSPathExp (..))
+import PostgREST.JSPath (JSPathExp (..))
 import Protolude        hiding (toS)
 import Protolude.Conv   (toS)
